+++ bigplanet/__init__.py
@@ -13,5 +13,5 @@ __version__ = "1.0"
 __author__ = "Caitlyn Wilhelm (cwilhelm@uw.edu)"
 __copyright__ = "Copyright 2020  Caitlyn Wilhelm"
 
-from .bigplanet_module import *
+from .bp_extract import *
 from .bigplanet import *
